+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-04-06 14:14:38
-// Commit: style(state): harmonize light and dark theme palette for state
+// Updated at: 2026-04-09 14:46:19
+// Commit: refactor(state): optimize data flow between domain and data layers
 
 class stateComponentState {
-  final String commitRef = '157';
-  final String lastMessage = 'style(state): harmonize light and dark theme palette for state';
-  final DateTime timestamp = DateTime.parse('2026-04-06T14:14:38+05:00');
+  final String commitRef = '226';
+  final String lastMessage = 'refactor(state): optimize data flow between domain and data layers';
+  final DateTime timestamp = DateTime.parse('2026-04-09T14:46:19+05:00');
 }